# packages/assets-service/openapi3-definition.json
{
  "openapi": "3.0.0",
  "paths": {
    "/assets": {
      "get": {
        "operationId": "AppController_listAssets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "assets"
        ]
      }
    },
    "/categories": {
      "get": {
        "operationId": "AppController_listAssetCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "assets"
        ]
      }
    },
    "/assets/{id}": {
      "get": {
        "operationId": "AppController_findOneAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "assets"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "AppController_findOneCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "assets"
        ]
      }
    }
  },
  "info": {
    "title": "Assets service example",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "assets",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "publisherId": {
            "type": "string",
            "example": "1"
          },
          "categoryId": {
            "type": "string",
            "example": "1"
          },
          "title": {
            "type": "string",
            "example": "MyAsset"
          }
        },
        "required": [
          "id",
          "publisherId",
          "categoryId",
          "title"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "3D"
          }
        },
        "required": [
          "id",
          "name"
        ]
      }
    }
  }
}